OP-5660 - refactored location of scripts
set_settings should be in lib as it is used elsewhere, but launch_logic and lib created circular dependency.
Moved main to launch logic as it is actually handling launching.
